Add files via upload
File contains the assignment of 1938 individual polar bears to their respective genetic cluster in Canada's Arctic along with percent missingness of each individual and probability of membership to each of the four genetic clusters based on the results of a STRUCTURE analysis outlined in the methods section of this thesis. Cluster 1 is the Hudson Bay Complex, Cluster 2 is the Arctic Archipelago, Cluster 3 is the Polar Basin, and Cluster 4 is M'Clintock Channel. Individuals were assigned to a cluster in two ways. Samples were assigned based on either the largest cluster they had membership to or the cluster that they had over 80% probability of membership to. Samples with less then 80% membership to any of the four clusters are denoted by NA in the final column.